onnx_output_exam.py
import onnx

model = onnx.load("/home/yolov11-3.0/modelv1/6.onnx")

print("输入:")
for inp in model.graph.input:
    print(f"  {inp.name}: {[d.dim_value for d in inp.type.tensor_type.shape.dim]}")

print("输出:")
for out in model.graph.output:
    print(f"  {out.name}: {[d.dim_value for d in out.type.tensor_type.shape.dim]}")
