requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-aarch64
# created-by: conda 25.9.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=51_gnu
absl-py=2.3.1=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.11.0=pypi_0
asttokens=3.0.1=pypi_0
attrs=25.4.0=pypi_0
bzip2=1.0.8=h998d150_6
ca-certificates=2025.11.4=hd43f75c_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.2=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.2.1=pypi_0
exceptiongroup=1.3.1=pypi_0
executing=2.2.1=pypi_0
expat=2.7.3=hc366171_0
fastapi=0.122.0=pypi_0
filelock=3.19.1=pypi_0
flatbuffers=25.9.23=pypi_0
fonttools=4.60.1=pypi_0
fsspec=2025.9.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
grpcio=1.76.0=pypi_0
grpcio-reflection=1.76.0=pypi_0
grpcio-tools=1.76.0=pypi_0
h11=0.16.0=pypi_0
humanfriendly=10.0=pypi_0
hyperlpr3=0.1.3=pypi_0
idna=3.4=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
ipython=8.18.1=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-aarch64=2.44=h4e4ac68_2
libffi=3.4.4=h419075a_1
libgcc=15.2.0=hc18542e_7
libgcc-ng=15.2.0=h51576c1_7
libgomp=15.2.0=hf47c802_7
libnsl=2.0.0=h998d150_0
libstdcxx=15.2.0=h9538471_7
libstdcxx-ng=15.2.0=h719ae8e_7
libuuid=1.41.5=h998d150_0
libxcb=1.17.0=hf66535e_0
libzlib=1.3.1=h998d150_0
loguru=0.7.3=pypi_0
markdown=3.9=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.2.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h419075a_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
onnxruntime=1.19.2=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openssl=3.0.18=h8a9c3cf_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
parso=0.8.5=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.3.0=pypi_0
pip=25.3=pyhc872135_0
prompt-toolkit=3.0.52=pypi_0
protobuf=6.33.1=pypi_0
psutil=7.1.3=pypi_0
pthread-stubs=0.3=hfd63f10_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pydantic=2.12.4=pypi_0
pydantic-core=2.41.5=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.2.5=pypi_0
python=3.9.25=h336a7db_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
readline=8.3=h886d1d0_0
requests=2.28.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=80.9.0=py39hd43f75c_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.51.0=hc5b20ee_0
stack-data=0.6.3=pypi_0
starlette=0.49.3=pypi_0
sympy=1.14.0=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.15=h987d8db_0
torch=2.1.0=pypi_0
torch-npu=2.1.0.post17=pypi_0
torchvision=0.16.0=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=1.26.13=pypi_0
uvicorn=0.38.0=pypi_0
wcwidth=0.2.14=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39hd43f75c_0
xorg-libx11=1.8.12=hf66535e_1
xorg-libxau=1.0.12=hf66535e_0
xorg-libxdmcp=1.1.5=hf66535e_0
xorg-xorgproto=2024.1=h998d150_1
xz=5.6.4=h998d150_1
zipp=3.23.0=pypi_0
zlib=1.3.1=h998d150_0
